requirement.txt
pytorch torchvision torchaudio cpuonly -c pytorch                                                                                                        (base) 
flask
pillow
requests